Adafruit IO bug fix
The sketch adafruit_io_digital_out does not work properly with Adafruit IO. The value returned by toPinLevel() is always 0 whether I turn on or off the LED.
The reason is that Adafruit IO toggle button feed sends ON and OFF values by default. The API doesn't work with those values so I added the necessary.