Compare feed names case-insensitively
Feeds are matched with a case-sensitive search when adding new feed
data to a Group.  This can cause different values to be published
for the same feed in a single call to save() in a way that's not
easy to see from the code.

Also, subscribing to a feed using the uppercase name of the feed
will never work since the feed names are lowercased when received
from AdafruitIO but the string compare here is strictly
case-sensitive.

Fix these problems by matching feed names case-insensitively when
searching for existing feeds in the Group data list and when matching
callback lookups.

Note: I added the new `strsame_nocase` function inline in the
AdafruitIO_Group.cpp module since I didn't see any clear place to put
these kinds of utility functions in the library.  I'm open to moving
it, renaming, restructuring it or whatever.  Perhaps a better option
is to use the `String::equalsIgnoreCase()`, but constructing String
temporaries for each comparison seemed excessive to me.

Fixes #123